Public vs internal API separation
Moved the concrete classes to internal implementation and exposed well documented interfaces
