+++ src/main/groovy/org/mockito/release/gradle/notes/ReleaseNotesExtension.java
@@ -1,0 +1,15 @@
+package org.mockito.release.gradle.notes;
+
+/**
+ * This extension object is added by {@link ReleaseNotesPlugin}. Example configuration:
+ *
+ * <pre>
+ *  notes {
+ *    notesFile = file("docs/release-notes.md")
+ *    gitHubAuthToken = "secret"
+ *    gitHubLabelMappings = [:]
+ *  }
+ * </pre>
+ */
+public interface ReleaseNotesExtension {
+}
