+++ src/main/groovy/org/mockito/release/gradle/notes/ReleaseNotesPlugin.java
@@ -1,9 +1,7 @@
 package org.mockito.release.gradle.notes;
 
-import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
-import org.gradle.api.Task;
 
 /**
  * The plugin adds following tasks:
@@ -13,7 +11,7 @@ import org.gradle.api.Task;
  *     <li>previewReleaseNotes - prints incremental release notes to the console for preview.</li>
  * </ul>
  *
- * The plugin also adds "notes" extension:
+ * The plugin also adds "notes" extension, for more information see {@link ReleaseNotesExtension}.
  *
  * <pre>
  *  notes {
@@ -23,34 +21,5 @@ import org.gradle.api.Task;
  *  }
  * </pre>
  */
-public class ReleaseNotesPlugin implements Plugin<Project> {
-
-    public void apply(Project project) {
-        final ReleaseNotesExtension notes = project.getExtensions().create(ReleaseNotesExtension.EXT_NAME, ReleaseNotesExtension.class,
-                project.getProjectDir(), project.getVersion().toString());
-        project.getTasks().create("updateReleaseNotes", new Action<Task>() {
-            public void execute(Task task) {
-                task.setGroup("Release Notes");
-                task.setDescription("Updates release notes file.");
-                task.doLast(new Action<Task>() {
-                    public void execute(Task task) {
-                        notes.updateReleaseNotes();
-                    }
-                });
-            }
-        });
-
-        project.getTasks().create("previewReleaseNotes", new Action<Task>() {
-            public void execute(Task task) {
-                task.setGroup("Release Notes");
-                task.setDescription("Shows new incremental content of release notes. Useful for previewing the release notes.");
-                task.doLast(new Action<Task>() {
-                    public void execute(Task task) {
-                        String content = notes.getReleaseNotes();
-                        task.getLogger().lifecycle("----------------\n$content----------------");
-                    }
-                });
-            }
-        });
-    }
+public interface ReleaseNotesPlugin extends Plugin<Project> {
 }
