+++ src/main/resources/META-INF/gradle-plugins/org.mockito.release-notes.properties
@@ -1,1 +1,1 @@
-implementation-class=org.mockito.release.gradle.notes.ReleaseNotesPlugin
+implementation-class=org.mockito.release.internal.gradle.DefaultReleaseNotesPlugin
